requirements.txt
# Google Drive API
PyDrive2>=1.16.0
PyYAML>=6.0

# Core RAG and LLM
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20
openai>=1.12.0

# Vector Database
chromadb>=0.4.22

# Document Processing
pypdf>=4.0.0
python-docx>=1.1.0

# Audio/Video Transcription
openai-whisper>=20231117
ffmpeg-python>=0.2.0

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
gunicorn>=21.2.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
tiktoken>=0.5.2
PyJWT>=2.8.0

